[bug]: fixed history not being deleted when leaving and refocusing tab
